requirements.txt
functions-framework==3.*
requests==2.32.3
loguru==0.7.3
google-cloud-bigquery==3.29.0
tenacity==9.0.0